feat(fonts): make FontStore font indexing page independent
In the current state, the font id is determined by the time of adding
the font. In the current state, the font with the id 0, is added first,
therefore it is located on page one.

To make parsing parts of document possible, determining the font id must
be independent from time and space, so hashing the font is the better
approach.

+ fix unit test
+ extend type interfaces